feat: created a separate form page and removed the form from the home page
