--- src/components/ConsultationSection.jsx
@@ -1,12 +1,0 @@
-import ConsultForm from "./ConsultForm"
-
-const ConsultationSection = () => {
-  return (
-    <div id="interestForm" className="consultation-container px-12">
-        <h4 className="text-5xl py-4 text-center">Interested in our services? Let's talk!</h4>
-        <ConsultForm />
-    </div>
-  )
-}
-
-export default ConsultationSection
