+++ src/pages/Home.jsx
@@ -1,6 +1,6 @@
 import Hero from "../components/Hero";
 import FoodMenuSection from "../components/FoodMenuSection";
-import ConsultationSection from "../components/ConsultationSection";
+// import ConsultationSection from "../components/ConsultationSection";
 import Services from "../components/Services";
 import Story from "../components/Story";
 import ReviewSection from "../components/ReviewSection";
@@ -12,7 +12,7 @@ const Home = () => {
       <Hero />
       <FoodMenuSection />
       <Services />
-      <ConsultationSection />
+      {/* <ConsultationSection /> */}
       <Story />
       <ReviewSection />
     </>
